7/8~7/12 @ GIS conference
This is made for the AP of 2019 GIS conference.
It is used to help users acknowledge financial knowledge as well as investment recommendation.

You can use different sentence with same meaning to talk to her.
Sample sentences:

I want to  invest.
I just want to learn.
What's Treasury bills?
You're smart.
Who are you?
I want to invest in real estates.

Note: This is created in 3 days, so it is not well-function.
      May have some bugs.